// src/store/Editor/mutation-type.ts
export enum MUTATION_TYPE {
  ADD_PAGE = "ADD_PAGE",
  REMOVE_PAGE = "REMOVE_PAGE",
  SELECT_PAGE = "SELECT_PAGE",
  ADD_COMPONENT = "ADD_COMPONENT",
  REMOVE_COMPONENT = "REMOVE_COMPONENT",
  SELECT_COMPONENT = "SELECT_COMPONENT",
  UNDO = "UNDO",
  REDO = "REDO",
  INIT = "INIT",
  UPDATE_COMPONENT = "UPDATE_COMPONENT",
  RESIZE = "RESIZE",
  ADD_EVENT = "ADD_EVENT",
  REMOVE_EVENT = "REMOVE_EVENT",
  UPDATE_EVENT = "UPDATE_EVENT",
  DRAG_COMPONENT = "DRAG_COMPONENT",
  DRAG_TREE = "DRAG_TREE",
}